chore: simplify memcache pipelining flow (#6230)
Before - memcache piggybacked on the RESP flow for pipelining.
It was overkill and overcomplicated as memcache does not perform command squashing,
and the only actual optimization that helped with increasing throughput for memcache pipelining was
the fact we batched replies via reply_builder_->SetBatch interface.

This PR rewrites the IO flow for memcache, by introducing an intrusive queue of parsed commands.
The first step is to parse one or many commands and the second
step is to execute them one by one. We enable batching of replies as long as we have more commands in queue.

The throughput of the new approach is similar to what we had before but we do not use dispatch fiber for this,
and we perform pipelining in batches of limited size.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>