chore: move memcache reply rendering into MCRender (#6267)
* chore: move memcache reply rendering into MCRender

We need to store deferred replies in the reply_payload, so we need to reuse
the same logic of rendering replies inside ParsedCommand. In order to preserve
the separation of responsibilities we introduce a helper class MCRender that
helps rendering replies for memcache commands.
The class is being used for writing replies directly into socket via ReplyBuilder, or
for caching them in reply_payload.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>

* chore: fixes

Signed-off-by: Roman Gershman <roman@dragonflydb.io>

---------

Signed-off-by: Roman Gershman <roman@dragonflydb.io>