chore: get rid of command specific flags in MCReplyBuilder
Before - we passed memcache command specific flags to MCReplyBuilder to remember the
response configuration. However, this does not work with asynchronous dispatch, as
a command can be replied at a differrent time from when these flags were set.
We need to use flags kept inside ParsedCommand to have consistent behavior.

Also, we can not render memcache replies via CapturingReplyBuilder as it is stateless
with respect to these flags, hence we move rendering of StoredReply response to ParsedCommand.

The other response that can not be rendered using CapturingReplyBuilder visitor is MGET response,
as it requires keys and memcache specific flags to render the response correctly. This logic will
be added in the subsequent PR.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>