core: Remove spurious check for should-defragment
Earlier the json defragmentation process was serialization based and a
one-shot operation. The decision to defrag was taken by first walking
the tree and figuring out if any node was in an under-utilized page. If
so the whole node was reallocated.

The new reallocation algorithm walks the tree and visits every node,
reallocating if the page is under-utilized. The first check is no longer
needed, it is now implicit in the defragmentation process.

Signed-off-by: Abhijat Malviya <abhijat@dragonflydb.io>