chore(server): Use cycle based quota for defragmentation (#6262)
* core: Add quota class which can be armed separate from ctor

The page usage class now accepts a quota instead of managing /
calculating this by itself. The quota can be explicitly armed to set the
start point.

* server: Switch from count based quota to cycles based quota

The defrag task used to run for 40 invocations. Now the task is split
into two phases. Each is allowed 50 microseconds to run. The first phase
defragments objects in db. The second phase defragments the search
indices.

* server: Remove nullptr check as follow up from prev. PR
* server/tests: Add test to ensure defragmentation eventually finishes

When running with a quota, defragmentation should eventually finish
after a few cycles.

To support the test, DoDefrag now accepts a page usage obj. reference,
so that it can be primed to force-reallocate every time.

Signed-off-by: Abhijat Malviya <abhijat@dragonflydb.io>