chore(json): Use new json defragmentation (#6276)
* core: Add a flag to optionall disable json defragmentation
* core: Use tree walk based json defragmentation
* core: Return bool from json defragment to convey info about operation
The boolean value will be used by callers to update stats and metrics.
* core: Remove spurious check for should-defragment
Earlier the json defragmentation process was serialization based and a
one-shot operation. The decision to defrag was taken by first walking
the tree and figuring out if any node was in an under-utilized page. If
so the whole node was reallocated.

The new reallocation algorithm walks the tree and visits every node,
reallocating if the page is under-utilized. The first check is no longer
needed, it is now implicit in the defragmentation process.

---------

Signed-off-by: Abhijat Malviya <abhijat@dragonflydb.io>