search: Use stateless allocator for strings
The strings in sort indices are changed to use stateless allocators
instead of polymorphic ones. The same change is also made to vectors.

The stateless allocator benifits from empty base optimization, no space
needs to be reserved for it within strings because it is an empty class.

A static assert is added to confirm this.

Signed-off-by: Abhijat Malviya <abhijat@dragonflydb.io>