search: Use a special stateless allocator bound to local mr
The local mr has a used bytes counter which only tracks usage for search
indices. In order to use a stateless allocator we must use one which
only works with this memory resource, so that the counts are not
inadvertently shared with the resource set up in engine shard.

To achieve this the stateless allocator from core/detail is copied and
uses its own thread local memory resource.

Signed-off-by: Abhijat Malviya <abhijat@dragonflydb.io>