chore: simplify GAT command arguments
Before - Dragonfly GAT command used `<TS> key1 .. keyn ` format to
pass expiry timestamp information further.
Now, we have memcache context as part of the ParsedCommand structure so we can access expire_ts directly.

This change moves the logic of parsing the correct timestamp to the parser and removes the first <TS> argument
from GAT. This will help with consolidating GAT and MGET code down the road.

This change is possible as we do not need to replicate GAT as is, as it is translated to PEXPIREAT or DEL commands.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>