chore: implement support for deferred mget
This does not implement async dispatching for MC/mget, just merely supports
replies in the correct order for cases where the previous command was dispatched
asynchronously. We need to implement it for every MC command to preserve the correctness.

While testing GET/GAT, I learned that GAT was broken for meta protocol so I added a few fixes
to be able to test it.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>