chore: introduce RespSrvParser
This is a limited, simpler and more convenient version of RESP parser, designed to
parse server-side expressions that look like a flat array of strings.

Motivation: we need a parser that parses arguments directly using BackedArgs so we could
use the parsed arguments for async dispatching.

I duplicated redis_parser code and stripped unneeded parts.
There some clean-ups that can be done in the subsequent PRs.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>